requirements.txt
absl-py==2.3.1
accelerate==1.10.1
addict==2.4.0
adjustText==1.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
argcomplete==3.6.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
autoflake==2.3.1
av==15.1.0
av2==0.2.1
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.14.3
black==25.11.0
bleach==6.2.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
cmake==3.25.0
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.0
cycler==0.12.1
datasets==4.5.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dependency-groups==1.3.1
descartes==1.1.0
dill==0.4.0
distlib==0.4.0
einops==0.8.1
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
fire==0.7.1
flake8==7.3.0
flash-attn==2.3.5
fonttools==4.60.2
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanize==4.13.0
idna==3.11
imageio==2.31.4
imageio-ffmpeg==0.4.9
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.31.0
ipython==8.12.2
ipywidgets==8.1.8
iso3166==2.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.7
lark==1.3.1
lit==15.0.7
llvmlite==0.36.0
lyft-dataset-sdk==0.0.8
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.2.0
mmcls==0.25.0
mmcv-full==1.7.0
mmdet==2.26.0
mmdet3d==1.0.0rc6
mmengine==0.10.7
mmsegmentation==0.29.1
model-index==0.1.11
modelscope==1.33.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.14.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.3
ninja==1.13.0
notebook==7.5.1
notebook_shim==0.2.4
nox==2025.11.12
numba==0.53.0
numpy==1.23.5
nuscenes-devkit==1.1.11
opencv-python==4.11.0.86
opendatalab==0.0.10
# Editable install with no version control (openlanev2==2.1.0)
-e ./OpenLane-V2-master
openmim==0.3.9
openxlab==0.0.11
ordered-set==4.1.0
ortools==9.3.10497
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
peft==0.17.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.4.0
plotly==6.5.0
pluggy==1.6.0
plyfile==1.1.3
prettytable==3.16.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycocotools==2.0.11
pycodestyle==2.14.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.3.1
pyproj==3.6.1
pyquaternion==0.9.9
pytest==8.4.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytokens==0.3.0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.3
pyzmq==27.1.0
rdp==0.8
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
safetensors==0.7.0
scikit-image==0.19.3
scikit-learn==1.6.1
scipy==1.8.0
Send2Trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.48.0
shapely==2.0.0
similaritymeasures==1.4.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.22
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.3.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.0
trimesh==2.35.39
triton==2.0.0
trl==0.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
virtualenv==20.35.4
wandb==0.23.1
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
widgetsnbextension==4.0.15
xxhash==3.6.0
yapf==0.40.1
yarl==1.22.0
zipp==3.23.0
